Handle non-string objects being added to properties files
This issue is most importantly caused by using string interpolation in
gradle, which creates a GString that is lazily interpolated when it is
used as a string somewhere. Explicitly cast all keys and values of the
properties maps to strings, so they are all included when a properties
file is written.

Fixes #11